+++ examples/shank/token-metadata/mpl_token_metadata_interface/src/typedefs.rs
@@ -1,10 +1,12 @@
 use borsh::{BorshDeserialize, BorshSerialize};
 use solana_program::pubkey::Pubkey;
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct MintPrintingTokensViaTokenArgs {
     pub supply: u64,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct SetReservationListArgs {
     pub reservations: Vec<Reservation>,
     pub total_reservation_spots: Option<u64>,
@@ -12,12 +14,14 @@ pub struct SetReservationListArgs {
     pub total_spot_offset: u64,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct UpdateMetadataAccountArgs {
     pub data: Option<Data>,
     pub update_authority: Option<Pubkey>,
     pub primary_sale_happened: Option<bool>,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct UpdateMetadataAccountArgsV2 {
     pub data: Option<DataV2>,
     pub update_authority: Option<Pubkey>,
@@ -25,32 +29,39 @@ pub struct UpdateMetadataAccountArgsV2 {
     pub is_mutable: Option<bool>,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct CreateMetadataAccountArgs {
     pub data: Data,
     pub is_mutable: bool,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct CreateMetadataAccountArgsV2 {
     pub data: DataV2,
     pub is_mutable: bool,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct CreateMasterEditionArgs {
     pub max_supply: Option<u64>,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct MintNewEditionFromMasterEditionViaTokenArgs {
     pub edition: u64,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct ApproveUseAuthorityArgs {
     pub number_of_uses: u64,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct UtilizeArgs {
     pub number_of_uses: u64,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct Data {
     pub name: String,
     pub symbol: String,
@@ -59,6 +70,7 @@ pub struct Data {
     pub creators: Option<Vec<Creator>>,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct DataV2 {
     pub name: String,
     pub symbol: String,
@@ -69,35 +81,41 @@ pub struct DataV2 {
     pub uses: Option<Uses>,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct Uses {
     pub use_method: UseMethod,
     pub remaining: u64,
     pub total: u64,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct Collection {
     pub verified: bool,
     pub key: Pubkey,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct Creator {
     pub address: Pubkey,
     pub verified: bool,
     pub share: u8,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct Reservation {
     pub address: Pubkey,
     pub spots_remaining: u64,
     pub total_spots: u64,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct ReservationV1 {
     pub address: Pubkey,
     pub spots_remaining: u8,
     pub total_spots: u8,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub enum Key {
     Uninitialized,
     EditionV1,
@@ -111,12 +129,14 @@ pub enum Key {
     CollectionAuthorityRecord,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub enum UseMethod {
     Burn,
     Multiple,
     Single,
 }
 #[derive(Clone, Debug, BorshDeserialize, BorshSerialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub enum TokenStandard {
     NonFungible,
     FungibleAsset,