+++ src/main.rs
@@ -81,11 +81,13 @@ pub struct Args {
         default_value = "^0.2"
     )]
     pub num_traits_vers: String,
-}
 
-fn check_valid_semver_req(arg: &str, arg_name: &str) {
-    semver::VersionReq::parse(arg)
-        .unwrap_or_else(|_| panic!("Invalid version req '{}' for {}", arg, arg_name));
+    #[arg(
+        long,
+        help = "serde dependency version for generated crate",
+        default_value = "^1.0"
+    )]
+    pub serde_vers: String,
 }
 
 fn main() {
@@ -96,9 +98,6 @@ fn main() {
     log_panics::init();
 
     let mut args = Args::parse();
-
-    check_valid_semver_req(&args.solana_program_vers, "solana-program");
-    check_valid_semver_req(&args.borsh_vers, "borsh");
 
     let mut file = OpenOptions::new().read(true).open(&args.idl_path).unwrap();