+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.4] - 2023-07-21
+
+### Added
+
+- `--serde-vers` to configure `serde` as an optional dependency for the generated crate
+
+### Changed
+
+- Allow toml maps to be passed to the various `--*-vers` options to allow for values like `"workspace = true"`
+
 ## [0.1.3] - 2023-06-19
 
 ### Changed